Compilation fix in tonemapwidget
VC++ does not allow conversion from floating point to enums. In any case, the m_false_method variable should be an integer, and the rest of the code involving it assumes that it is.

+++ qtgui/tonemapwidget.hxx
@@ -85,7 +85,7 @@ private:
 	double m_TM_contrast_ywa;
 
 	int m_false_method;
-	double m_false_colorscale;
+	int m_false_colorscale;
 	double m_TM_FalseMax;
 	double m_TM_FalseMin;
 	double m_TM_FalseMaxSat;
